move renderPrivateRoute
